% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absolutelyWright.r
\name{absolutelyWright}
\alias{absolutelyWright}
\title{Function absolutelyWright}
\usage{
absolutelyWright(dat1, report.alpha = T, main.title = "Wright Map",
  axis.persons = "Persons", axis.items = "Items",
  axis.logits = "Logits", item.prop = 0.5, show.axis.logits = "R",
  item.side = WrightMap::itemClassic, dim.color = "#C6DBEF", ...)
}
\description{
Plots Wright Map based on data.
}
