% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkItemQuality.r
\name{checkItemQuality}
\alias{checkItemQuality}
\title{Function checkItemQuality}
\usage{
checkItemQuality(data, exclude = NULL, cutoffs = c(0.2, 2, 2, 0.3),
  reduce = T)
}
\description{
Returns information on item fit based on data.
}
