% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkModelQuality.r
\name{checkModelQuality}
\alias{checkModelQuality}
\title{Function checkModelQuality}
\usage{
checkModelQuality(data)
}
\description{
Returns information on model fit and psychomeric quality based on data.
}
