% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCCC.r
\name{plotCCC}
\alias{plotCCC}
\title{Function plotCCC}
\usage{
plotCCC(difficulties, diff.names = NULL,
  main = "Category Characteristic Curves", labeling = T, lwd = 2,
  xlim = NULL, col = NULL, ...)
}
\description{
Plots Category Characteristic Curves based on a named vector of item difficulties.
}
