% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotICC.r
\name{plotICC}
\alias{plotICC}
\title{Function plotICC}
\usage{
plotICC(difficulties, slopes = seq(1:length(difficulties)) * 0 + 1,
  main = "Item Characteristic Curves", labeling = T, xlim = NULL,
  lwd = 2, col = NULL)
}
\description{
Plots Item Characteristic Curves based on a vector of item difficulties.
}
