% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.data.r
\name{random.data}
\alias{random.data}
\title{Function random.data}
\usage{
random.data(simulatedPersons = 100, simulatedItems = 10,
  addMissings = 10)
}
\description{
Generates random data based on the Rasch model.
}
