% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkItemQuality.r
\name{checkItemQuality}
\alias{checkItemQuality}
\title{Function checkItemQuality}
\usage{
checkItemQuality(data, deleted = NULL)
}
\description{
Returns information on item fit based on data.
}
